fastapi/requirements.txt
numpy==1.24.4
pandas==2.0.3
fastapi==0.110.0
uvicorn==0.29.0
psycopg2-binary==2.9.9
requests==2.31.0

